% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mean_min_abs_distance}
\alias{mean_min_abs_distance}
\title{Given an array of voter positions and party positions,
computes the mean of the minimum absolute distances from each voter to the nearest party.}
\usage{
mean_min_abs_distance(voters, parties)
}
\description{
Given an array of voter positions and party positions,
computes the mean of the minimum absolute distances from each voter to the nearest party.
}
