% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mean_min_abs_distance_mc_batch}
\alias{mean_min_abs_distance_mc_batch}
\title{Monte Carlo simulation entirely in Rust for fair benchmarking
Runs multiple simulations and returns the mean result}
\usage{
mean_min_abs_distance_mc_batch(parties, n_voters, n_simulations, seed)
}
\description{
Monte Carlo simulation entirely in Rust for fair benchmarking
Runs multiple simulations and returns the mean result
}
