% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mean_min_abs_distance_midpoints}
\alias{mean_min_abs_distance_midpoints}
\title{Midpoint-based approach to computing expected distance to nearest party}
\usage{
mean_min_abs_distance_midpoints(parties)
}
\description{
This divides the voter space into segments between party midpoints,
then integrates the expected distance to the nearest party in each segment.
}
